// src/app/models/registration.ts
export interface Registration {
    id?: number;
    name?: string;
    userName?: string;
    userType?: number;
    email?: string;
    password?: string;
    contact?: string;
    address?: string;
    pinCode?: number;
    cityId?: number;
    stateId?: number;
    city?: string;
    state?: string;
    result?:string;  
}